Update and rename execute-workflow.ts to start-workflow.ts
